+++ GeneralsMD/Code/GameEngine/Include/Common/UserPreferences.h
@@ -112,6 +112,7 @@ public:
 	Bool getSnapBuildPlacementTo45(void);
 	Bool getSnapCameraRotateTo45(void);
 	Bool getGridBuildPlacement(void);
+	Bool getNudgeBuildPlacement(void);
 	Bool getMiddleMousePans(void);
 	Bool getZoomToCursor(void);
 	Bool getShowHudOverlay(void);